Add files via upload
Title: Airbnb Clone - Replicating the Vacation Rental Platform
In this project, I've created an Airbnb clone file using HTML and Tailwind CSS. The aim is to replicate the sleek and modern design of the popular vacation rental platform, Airbnb. By leveraging HTML for structuring and Tailwind CSS for styling, this clone file mimics the user interface and layout of Airbnb's website.



Project Status:
The Airbnb Clone file created with HTML and Tailwind CSS serves as a foundation for replicating the user interface and design aesthetics of the Airbnb platform. Further development may involve integrating backend functionality, implementing interactivity with JavaScript, and refining the visual design to closely resemble the original website.